feat(cli): retrieve version from IntegrationPlatform
As mentioned in #1912, moving the logic to recover the operator version from each namespace IntegrationPlatform. This would solve the problem when any global operator is installed.

Ref #1652
