+++ pkg/cmd/util.go
@@ -43,8 +43,7 @@ import (
 )
 
 const (
-	offlineCommandLabel                     = "camel.apache.org/cmd.offline"
-	versionCompatibilityWarningCommandLabel = "camel.apache.org/cmd.versionCompatibilityWarning"
+	offlineCommandLabel = "camel.apache.org/cmd.offline"
 )
 
 // DeleteIntegration --
@@ -200,10 +199,6 @@ func cmdOnly(cmd *cobra.Command, options interface{}) *cobra.Command {
 
 func isOfflineCommand(cmd *cobra.Command) bool {
 	return cmd.Annotations[offlineCommandLabel] == "true"
-}
-
-func showCompatibilityWarning(cmd *cobra.Command) bool {
-	return cmd.Annotations[versionCompatibilityWarningCommandLabel] == "true"
 }
 
 func clone(dst interface{}, src interface{}) error {
